Move ConditionalWrapper to /components
