+++ packages/manager/src/components/ConditionalWrapper/ConditionalWrapper.tsx
@@ -23,3 +23,5 @@ export const ConditionalWrapper: React.FC<Props> = props => {
   const { condition, wrapper, children } = props;
   return condition ? wrapper(children) : <>{children}</>;
 };
+
+export default ConditionalWrapper;
