+++ packages/manager/src/components/ConditionalWrapper/index.ts
@@ -1,0 +1,2 @@
+import ConditionalWrapper from './ConditionalWrapper';
+export default ConditionalWrapper;
